% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addClusterGroup}
\alias{addClusterGroup}
\title{Assign group.ID for similar cluster.IDs}
\usage{
addClusterGroup(cluster.gr = NULL, cluster.groups = NULL)
}
\arguments{
\item{cluster.gr}{A \code{\link{GRanges-class}} object with contig position and their cluster assignment in 'clust.ID' metacolumn.}

\item{cluster.groups}{A \code{list} of grouped 'clust.IDs'.}
}
\value{
A \code{\link{GRanges-class}} object with an extra metacolumn group.ID.
}
\description{
This function will write extra metacolumn 'group.ID' into an input \code{\link{GRanges-class}} object
based on cluster.groups parameter.
}
\author{
David Porubsky
}
