% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{clusters2ranges}
\alias{clusters2ranges}
\title{Convert saarclust object into \code{\link{GRanges-class}} object.}
\usage{
clusters2ranges(saarclust.obj = NULL, best.prob = 1, prob.th = 0.5)
}
\arguments{
\item{saarclust.obj}{A path to the minimap file to load.}

\item{best.prob}{Set number of highest probability clusters reported per genomic region.}

\item{prob.th}{Only regions with this probability and higher will be reported.}
}
\value{
A \code{\link{GRanges-class}} object with reported cluster ID per genomic region.
}
\description{
This function takes saarclust object and for each region reports user defined number of clusters 
each regions likely belongs as a \code{\link{GRanges-class}} object.
}
\author{
David Porubsky
}
