% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseBins}
\alias{collapseBins}
\title{Collapse consecutive bins with the same ID value}
\usage{
collapseBins(gr, id.field = 0, measure.field = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}

\item{id.field}{A number of metadata column to use for region merging.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
Collapse consecutive bins with the same value defined in 'id.field'.
}
\author{
David Porubsky
}
