% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusterPartners_deprecated.R
\name{connectDividedClusters2}
\alias{connectDividedClusters2}
\title{Get pairs of clusters likely coming from the same chromosome}
\usage{
connectDividedClusters2(theta.param = NULL, z.limit = 2.58,
  remove.always.WC = TRUE)
}
\arguments{
\item{theta.param}{A \code{list} of estimated cell types for each cluster and each cell.}

\item{z.limit}{Connect clusters with z-score equal or above this limit.}
}
\value{
A \code{matrix} of pairs of clusters IDs that belong to the same chromosome.
}
\description{
This function finds cluster likely coming from the same chromosome by looking for clusters that share the same
directionality across multiple Strand-seq libraries.
}
\author{
David Porubsky
}
