% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDirectionalReads.R
\name{countDirectionalReads}
\alias{countDirectionalReads}
\title{Count directional reads.}
\usage{
countDirectionalReads(tab.l = NULL)
}
\arguments{
\item{tab.l}{A \code{list} of PB alignmetns separated per cell.}
}
\value{
A \code{list} of
}
\description{
This function expects output from custom minimap test dataset that contains original locations of mapped reads in the genome.
}
\author{
David Porubsky
}
