% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProbs.R
\name{countProb}
\alias{countProb}
\title{Calculate probability of PacBio reads having WW, CC or WC state.}
\usage{
countProb(minusCounts, plusCounts)
}
\arguments{
\item{minusCounts}{Minus (Watson) read counts aligned to PacBio reads.}

\item{plusCounts}{Plus (Crick) read counts aligned to PacBio reads.}
}
\value{
A \code{list} of binomial probabilities for a given counts of plus and minus reads
}
\description{
Import counts of directional Strand-seq reads aligned to PacBio reads and calculates binomial probabilites of PacBio reads being WW, CC or WC state.
}
\author{
David Porubsky, Maryam Ghareghani
}

