% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{counts2ranges}
\alias{counts2ranges}
\title{Report W & C counts and cluster ID as \code{\link{GRanges-class}} object.}
\usage{
counts2ranges(counts.l = NULL, saarclust.obj = NULL, best.prob = 1,
  prob.th = 0.5)
}
\arguments{
\item{counts.l}{A \code{list} of plus and minus alignments per genomic region.}

\item{saarclust.obj}{A path to the minimap file to load.}

\item{best.prob}{Set number of highest probability clusters reported per genomic region.}

\item{prob.th}{Only regions with this probability and higher will be reported.}
}
\value{
A \code{\link{GRangesList-class}} object with reported cluster ID per genomic region.
}
\description{
This function takes W and C counts togther cluster probabilties and reports user defined number of cluster IDs
for each genomic region and every single cell as a \code{\link{GRangesList-class}} object.
}
\author{
David Porubsky
}
