% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardClust.R
\name{estimateTheta}
\alias{estimateTheta}
\title{Estimate theta values based on hard clustering}
\usage{
estimateTheta(counts.l = NULL, hard.clust = NULL, alpha = 0.1)
}
\arguments{
\item{counts.l}{A \code{list} of directional read counts per PB read per library.}

\item{hard.clust}{A \code{integer} of cluster assignments for each PacBio read.}

\item{alpha}{Estimated level of background Strand-seq reads.}
}
\value{
A \code{list} of estimated theta values for every cluster and cell.
}
\description{
This function takes results of hard clustering and estimates majority cell types for each Strand-seq library
}
\author{
David Porubsky
}
