% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{exportClusteredReads}
\alias{exportClusteredReads}
\title{Export corresponding clusters for true chromosome}
\usage{
exportClusteredReads(inputfolder = NULL, prob.th = NULL, minLib = NULL)
}
\arguments{
\item{inputfolder}{Path to the data analysis folder}

\item{prob.th}{Filter out long reads with max probability below this threshold}

\item{minLib}{Filter out long reads with number of StrandS libraries being represented below this threshold}
}
\description{
Export corresponding clusters for true chromosome
}
\author{
David Porubsky
}
