% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{exportGenomicLocations}
\alias{exportGenomicLocations}
\title{Export best cluster IDs for each PB read}
\usage{
exportGenomicLocations(soft.clust, prob.th = 0.6)
}
\arguments{
\item{soft.clust}{...}

\item{prob.th}{...}
}
\description{
Function to get best clust ID for each PB read or eventually multiple cluster IDs if max pVal is lower than threshold
}
\author{
David Porubsky
}
