% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPseudoChromosomalScaffolds.R
\name{exportPseudoChromosomalScaffolds}
\alias{exportPseudoChromosomalScaffolds}
\title{Concatenate FASTA sequences given a set of Genomic Ranges and cluster identity.}
\usage{
exportPseudoChromosomalScaffolds(
  clustered.gr = NULL,
  assembly.fasta = NULL,
  outputfolder = NULL,
  concat.fasta = TRUE
)
}
\arguments{
\item{clustered.gr}{A \code{\link{GRanges-class}} object with given fields: dir - contig directionality
order - contig order and ID - contig assignment to clusters.}

\item{assembly.fasta}{An assembly FASTA file to extract DNA sequence determined by clustered.gr.}

\item{outputfolder}{A path to a folder to export concatenated FASTA sequences.}

\item{concat.fasta}{Set to \code{TRUE} if you want to concatenate FASTA sequences within a cluster by 100 N's.}
}
\description{
This function takes a \code{\link{GRanges-class}} object and cluster assignment stored 
in 'ID' field and export concated FASTA file for each cluster and genomic region/contig 
assigned to a given cluster.
}
\author{
David Porubsky
}
