% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{filterInput}
\alias{filterInput}
\title{Filter input data from the minimap}
\usage{
filterInput(inputData = NULL, quantileSSreads = c(0.4, 0.9),
  minSSlibs = 20)
}
\arguments{
\item{inputData}{A \code{data.frame} loaded by \code{\link[SaaRclust]{importTestData}}}

\item{quantileSSreads}{A quantile range for number of SSreads mapped to PB read. (default: 0.4-0.9)}
}
\value{
A filtered \code{data.frame}.
}
\description{
This function filters loaded data from the minimap.
}
\author{
David Porubsky
}
