% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusterPartners.R
\name{findAntiparallelClusters}
\alias{findAntiparallelClusters}
\title{Get pairs of clusters likely coming from the same chromosome but differ in directionality}
\usage{
findAntiparallelClusters(theta.param = NULL, z.limit = 2.58)
}
\arguments{
\item{theta.param}{A \code{list} of estimated cell types for each cluster and each cell.}
}
\value{
A \code{matrix} of pairs of clusters IDs that belong to the same chromosome.
}
\description{
This function finds cluster likely coming from the same chromosome by looking for clusters that have opposite
directionality across multiple Strand-seq libraries.
}
\author{
David Porubsky
}
