% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusterPartners.R
\name{findClusterPartners}
\alias{findClusterPartners}
\title{Get pairs of clusters coming from the same chromosome that differs in directionality.}
\usage{
findClusterPartners(theta.param = NULL)
}
\arguments{
\item{theta.param}{A \code{list} of estimated cell types for each cluster and each cell.}
}
\value{
A \code{vector} of pairs of clusters IDs that belong to the same chromosome.
}
\description{
This function solves the Maximum matching problem for all possible pairs of clusters and reports pairs with highest similarity.
}
\author{
David Porubsky, Maryam Ghareghani
}
