% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumped_functions.R
\name{findClusterPartners2}
\alias{findClusterPartners2}
\title{Get pairs of clusters coming from the same chromosome with different directionalities}
\usage{
findClusterPartners2(theta.param = NULL)
}
\arguments{
\item{theta.param}{A \code{list} containing estimated cell types per cluster for each cell.}
}
\value{
A \code{vector} of pairs of clusters IDs that belong to the same chromosome.
}
\description{
This function solves the Maximum matching problem for all possible pairs of clusters and reports pairs with highest similarity.
}
\author{
Maryam Ghareghani, David Porubsky
}
