% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumped_functions.R
\name{findClusterPartners_simple}
\alias{findClusterPartners_simple}
\title{Get pairs of clusters coming from the same chromosome with different directionalities}
\usage{
findClusterPartners_simple(theta.param = NULL)
}
\arguments{
\item{theta.param}{A \code{list} containing estimated cell types per cluster for each cell.}
}
\value{
A \code{vector} of pairs of clusters IDs that belong to the same chromosome.
}
\description{
This function finds the pairs by computing the pairwise distances between clusters theta_wc parameters, and pais each cluster with the cluster with min dist
Note that the first 23 minimum values should form a matching in the complete graph of clusters (this condition is observed to be held in practice)
TODO: We should replace this simple function by a proper max matching algorithm or some other solution that properly solves this problem and finds non-overlappinh pairs of clusters
}
\author{
Maryam Ghareghani
}
