% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusterPartners.R
\name{findSplitedClusters}
\alias{findSplitedClusters}
\title{Get pairs of clusters coming from the same chromosome}
\usage{
findSplitedClusters(theta.param = NULL)
}
\arguments{
\item{theta.param}{A \code{list} of estimated cell types for each cluster and each cell.}
}
\value{
A \code{vector} of pairs of clusters IDs that belong to the same chromosome.
}
\description{
This function finds cluster coming from the same chromosome and having the same directionality
}
\author{
David Porubsky
}
