% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flipCounts}
\alias{flipCounts}
\title{Reverse orientation of directional read counts}
\usage{
flipCounts(counts.l)
}
\arguments{
\item{counts.l}{A \code{list} of plus and minus alignments per genomic region.}
}
\value{
A \code{list} of matrices (columns: minus (W) and plus (C) counts; rows: genomic regions).
}
\description{
This function flips the orientation of directional reads counts for each genomic position.
}
