% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProbs.R
\name{gammaFunction}
\alias{gammaFunction}
\title{Calculate gamma function of EM algorithm}
\usage{
gammaFunction(clust.prob = NULL, pi = NULL)
}
\arguments{
\item{clust.prob}{A \code{list} of binomial probabilities of each PB read in every possible cluster}

\item{pi}{Paramter estimate of the size of each cluster}
}
\value{
A \code{list} of ...
}
\description{
Takes binomial probabilites for every cluster ...
}
\author{
David Porubsky, Maryam Ghareghani
}

