% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{getClusterAcc}
\alias{getClusterAcc}
\title{Check clustering accuracy}
\usage{
getClusterAcc(clusters)
}
\arguments{
\item{clusters}{List of clusters with predicted chromosome identity for each PB read}
}
\description{
Takes assigned identity of each PB read and compares it to known location.
}
\author{
David Porubsky
}

