% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{getClusterIdentity}
\alias{getClusterIdentity}
\title{Export corresponding clusters for each chromosome}
\usage{
getClusterIdentity(soft.clust, chr.rows)
}
\arguments{
\item{soft.clust}{...}

\item{chr.rows}{...}
}
\description{
Export corresponding clusters for each chromosome
}
\author{
David Porubsky
}

