% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{getClusterIdentityPerChr}
\alias{getClusterIdentityPerChr}
\title{Export corresponding clusters for true chromosome}
\usage{
getClusterIdentityPerChr(soft.clust, chr.rows)
}
\arguments{
\item{soft.clust}{Soft clustering probabilities for each long read and each cluster}

\item{chr.rows}{Expected location of long reads based on their mapping to a respective chromosome}
}
\description{
Export corresponding clusters for true chromosome
}
\author{
David Porubsky
}

