% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{getClusterIdentityPerChrPerDir}
\alias{getClusterIdentityPerChrPerDir}
\title{Export corresponding clusters for each chromosome}
\usage{
getClusterIdentityPerChrPerDir(soft.clust, chr.rows, chr.flag)
}
\arguments{
\item{soft.clust}{...}

\item{chr.rows}{...}

\item{chr.flag}{...}
}
\description{
Export corresponding clusters for each chromosome
}
\author{
David Porubsky
}

