% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getClusteringAcc}
\alias{getClusteringAcc}
\title{Report clustering accuracy of de novo assembled contigs}
\usage{
getClusteringAcc(bed.data, cluster.ID = "ID", contig.ID = "ctg")
}
\arguments{
\item{bed.data}{A \code{data.frame} object containing alignments of de novo assembly to the reference genome in BED format.}

\item{cluster.ID}{A column name the uniquely defines assignment of contigs to unique clusters.}

\item{contig.ID}{A column name the uniquely defines contigs aligned to the reference genome.}
}
\value{
A \code{data.frame} object.
}
\description{
This function takes alignments of de novo assembled contigs to the referene genome in BED format and reports
accuracy of their assignment to unique clusters.
}
\author{
David Porubsky
}
