% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCoinheritanceMatrix}
\alias{getCoinheritanceMatrix}
\title{Function to obtain co-inheritance matrix based on shared strand states}
\usage{
getCoinheritanceMatrix(contig.states = NULL, scale = FALSE,
  WWtoCC.penalty = TRUE)
}
\arguments{
\item{contig.states}{A \code{data.frame} of strand states per contig and per cell.}

\item{scale}{If set to \code{TRUE} returned values will be scaled to 1.}

\item{WWtoCC.penalty}{Penalize cases where there is W and C state in a single cell}
}
\value{
A symetric \code{matrix} of similarities for each contig pair.
}
\description{
This function takes \code{data.frame} of strand states per contig [rows] and per cell [columns]
and converts it into a matrix of contig similarities based on co-inheritance patterns.
}
\author{
David Porubsky
}
