% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{getQualMeasure}
\alias{getQualMeasure}
\title{Obtain summary measures of minimap output.}
\usage{
getQualMeasure(data.tab)
}
\arguments{
\item{data.tab}{Imported data table.}
}
\description{
Takes imported data table and export relevant data quality measures.
}
\author{
David Porubsky
}
