% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getRepresentativeAlignments}
\alias{getRepresentativeAlignments}
\title{Import representative PacBio alignments.}
\usage{
getRepresentativeAlignments(inputfolder = NULL, numAlignments = 50000)
}
\arguments{
\item{inputfolder}{A folder name where minimap files are stored.}

\item{numAlignments}{A required number of representative PacBio alignments.}

\item{store}{...}
}
\value{
A \code{data.frame}.
}
\description{
This function selects the best representative PacBio alignements from multiple chunks
in order to get the best estimates of theta values using Kmeans.
}
\author{
David Porubsky
}

