% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{hardClustAccuracy}
\alias{hardClustAccuracy}
\title{Computes the performance of the hard clustering algorithm}
\usage{
hardClustAccuracy(hard.clust, pb.chr, pb.flag, tab.filt, female = TRUE)
}
\arguments{
\item{hard.clust}{An \code{integer} vecotr of cluster assignements of each PacBio read.}

\item{pb.chr}{The true chromosomes of PacBio reads.}

\item{pb.flag}{The true directionality of PacBio reads.}

\item{tab.filt}{A \code{data.frame} obejct containing selected the best PacBio alignments.}

\item{female}{Set to \code{TRUE} if analyzed data are coming from female individual.}
}
\description{
This function takes hard clustering results and check the accuracy against know genomic locations of PacBio reads
}
\author{
Maryam Ghareghani, David Porubsky
}
