% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBams.R
\name{importBams}
\alias{importBams}
\title{Import BAM(s) and count reads}
\usage{
importBams(bamfolder = bamfolder, chromosomes = NULL,
  pairedEndReads = TRUE, min.mapq = 10, bin.size = 1e+05,
  step.size = NULL, reads.per.bin = NULL, max.frag = 1000,
  bin.method = "fixed", blacklist = NULL)
}
\arguments{
\item{bamfolder}{A folder containing BAM files with Strand-seq reads aligned to denovo assembly.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{bin.size}{A length of a bin to count reads in.}

\item{reads.per.bin}{An approximate number of desired reads per bin. The bin size will be selected accordingly. Forces 'bin.method' to be 'fixed'.}

\item{max.frag}{A maximum fragment length to import from the BAM file.}

\item{bin.method}{One of the 'fixed' or 'dynamic' binning method.}

\item{blacklist}{A \code{\link[GenomicRanges]{GRanges}} object containing genonic regions that should be excluded from the analysis.}
}
\value{
A \code{list} of matrices (columns: minus (W) and plus (C) counts; rows: genomic regions).
}
\description{
Import aligned reads from a multiple BAM files and counts directional reads in specified genomic locations.
Results are stored in a \code{list} of matrices with each element of a \code{list} representing counts for single BAM file.
}
\author{
David Porubsky
}
