% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importData}
\alias{importData}
\title{Import pre-processed output from the minimap}
\usage{
importData(infile = NULL, removeDuplicates = TRUE)
}
\arguments{
\item{infile}{A path to the minimap file to load.}
}
\value{
A \code{data.frame}.
}
\description{
This function expects output from custom minimap test dataset that contains original locations of mapped reads in the genome.
}
\author{
David Porubsky
}

