% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumped_functions.R
\name{importOldTestData}
\alias{importOldTestData}
\title{Import old output from the minimap}
\usage{
importOldTestData(infile = NULL, removeDuplicates = TRUE)
}
\arguments{
\item{infile}{A path to the minimap file to load.}
}
\value{
A \code{data.frame}.
}
\description{
This function expects output from custom minimap test dataset that contains original locations of mapped reads in the genome.
}
\author{
David Porubsky
}
