% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importTestData}
\alias{importTestData}
\title{Import output from the minimap}
\usage{
importTestData(infile = NULL, removeDuplicates = TRUE)
}
\arguments{
\item{infile}{A path to the minimap file to load.}

\item{removeDuplicates}{Set to \code{TRUE} if you want to remove duplicate reads based on the flag.}
}
\value{
A \code{data.frame}.
}
\description{
This function expects output from custom minimap test dataset that contains original locations of mapped reads in the genome.
}
\author{
David Porubsky
}
