% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{kahansum}
\alias{kahansum}
\title{Alternative sum function to reduce numerical error caused by finite precision of floating point numbers.}
\usage{
kahansum(x)
}
\arguments{
\item{x}{A \code{vector} of values tp sum up.}
}
\description{
Alternative sum function to reduce numerical error caused by finite precision of floating point numbers.
}
