% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelGenomicRegions}
\alias{labelGenomicRegions}
\title{Subtract and label set of genomic ranges in other set of genomic ranges.}
\usage{
labelGenomicRegions(gr = NULL, label.gr = NULL, label.gr.ID = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object to be labeled based on the overlap with 'label.gr' object.}

\item{label.gr}{A \code{\link{GRanges-class}} object of genomic regions to be labeled inside 'gr' object.}

\item{label.gr.ID}{A user defined label to be used to mark 'label.gr' regions inside 'gr'.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function takes a \code{\link{GRanges-class}} object of genomic regions to be labeled by user defined label
 in other \code{\link{GRanges-class}} object.
}
\author{
David Porubsky
}
