% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBins.R
\name{makeFixedBins}
\alias{makeFixedBins}
\title{Make fixed-width bins}
\usage{
makeFixedBins(bamfile = NULL, bin.size = 1e+05, step.size = NULL,
  chromosomes = NULL, keep.small.chr = FALSE)
}
\arguments{
\item{bamfile}{A BAM file from which the header is read to determine the chromosome lengths. If a \code{bamfile} is specified, option \code{assembly} is ignored.}

\item{bin.size}{A size of the genomic bin to split genome into.}

\item{step.size}{A size of the genomic interval to move each bin.}

\item{chromosomes}{A user defined set of chromosomes for binning (eg. 'chr1')}

\item{keep.small.chr}{Set to \code{TRUE} if chromosome/contigs smaller than the 'bin.size' should be kept.}
}
\value{
A \code{\link{GRanges-class}} object with fixed-width bins.
}
\description{
Make fixed-width bins based on given bin and step size.
}
\examples{
## Get an example BAM (empty bam)
bamfile <- system.file("extdata", "example.bam", package = "SaaRclust")
## Prepare genomic bins
genomic.bins <- makeFixedBins(bamfile = bamfile, bin.size = 200000, step.size = 200000, chromosomes = 'chr1')

}
\author{
David Porubsky
}
