% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskAlwaysWCandZeroBins.R
\name{maskAlwaysWCandZeroBins}
\alias{maskAlwaysWCandZeroBins}
\title{This function takes list of BAM files and enumerates user defined genomic bins 
for recurrent regions of WC strand-state and low read counts. Such bins are reported
as a custom \code{\link[GenomicRanges]{GRanges}} object or regions to be blacklisted.}
\usage{
maskAlwaysWCandZeroBins(
  bamfolder = bamfolder,
  genomic.bins = NULL,
  min.mapq = 10,
  pairedEndReads = TRUE,
  max.frag = 1000
)
}
\arguments{
\item{bamfolder}{A folder containing BAM files with Strand-seq reads aligned to denovo assembly.}

\item{genomic.bins}{A \code{\link[GenomicRanges]{GRanges}} object with defined genomic bins.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{max.frag}{A maximum fragment length to import from the BAM file.}
}
\description{
This function takes list of BAM files and enumerates user defined genomic bins 
for recurrent regions of WC strand-state and low read counts. Such bins are reported
as a custom \code{\link[GenomicRanges]{GRanges}} object or regions to be blacklisted.
}
\author{
David Porubsky
}
