% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardClust.R
\name{mergeClusters}
\alias{mergeClusters}
\title{Hierarchical clustering for merging the kmeans clusters.}
\usage{
mergeClusters(hard.clust, theta.l, k = 46)
}
\arguments{
\item{hard.clust}{The kmeans hard clustering.}

\item{theta.l}{A \code{list} of estimated theta values for each cluster and cell.}

\item{k}{Desired number of clusters after merging.}
}
\value{
A new hard clustering with the correct number of clusters
}
\description{
This function takes as input the kmeans hard clustering output and the initialized thetas and merges the kmeans clusters based on thetas
}
\author{
Maryam Ghareghani
}
