% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusterPartners.R
\name{mergeSplitedClusters}
\alias{mergeSplitedClusters}
\title{Merge splitted clusters}
\usage{
mergeSplitedClusters(cluster2merge = NULL, soft.pVal = NULL)
}
\arguments{
\item{cluster2merge}{...}

\item{soft.pVal}{...}
}
\description{
This function merges initialy divided clusters that belongs to the same chromosome.
}
\author{
David Porubsky
}
