% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderAndOrientClusters.R
\name{orderAndOrientClusters}
\alias{orderAndOrientClusters}
\title{Order and orient contigs within each cluster}
\usage{
orderAndOrientClusters(clustered.grl, split.pairs, ord.method = "TSP",
  alpha = 0.1, min.region.to.order = 0, filename = NULL,
  remove.always.WC = FALSE)
}
\arguments{
\item{clustered.grl}{A \code{\link{GRangesList-class}} object with cluster assignments for each contig.}

\item{split.pairs}{A \code{list} with a sets of cluster IDs that belong the same scaffold/chromosome.}

\item{ord.method}{A method ('TSP' or 'greedy') used to order contigs within a cluster (default: 'TSP')}

\item{alpha}{Estimated level of background in Strand-seq reads.}

\item{min.region.to.order}{A minimum size region to try to order within a cluster.}

\item{filename}{A path to a file where clustered on ordered contigs should be stored.}

\item{remove.always.WC}{Set to \code{TRUE} if the cluster with majority of WC states should be removed.}
}
\description{
This function takes a \code{\link{GRangesList-class}} object and order and orient each contigs
within the cluster and reports new clustered/scaffolded FASTA file.
}
\examples{
\dontrun{
## Get example files
mergeClusters.file <- system.file("extdata", "connectedClusters_5e+06bp_dynamic.RData", package = "SaaRclust")
## Load BAM count table
## Load connected clusters
connected.clusters <- get(load(mergeClusters.file))
ordered.contigs.gr <- orderAndOrientClusters(clustered.grl=clustered.grl, split.pairs=connected.clusters, min.region.to.order=0)
} 
}
\author{
David Porubsky
}
