% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderContigs.R
\name{orderContigsGreedy}
\alias{orderContigsGreedy}
\title{Order contigs based on shared strand states}
\usage{
orderContigsGreedy(contig.states, randomAttempts = 1000)
}
\arguments{
\item{contig.states}{A \code{data.frame} of strand states per contig and per cell.}

\item{randomAttempts}{Number of random trials to solve contig ordering.}
}
\description{
This function uses ContiBAIT implementation of MC algorithm to solve TSP
Copyright (c) 2015, Kieran O'Neill, Mark Hills, Mike Gottlieb
All rights reserved.
}
\author{
Kieran O'Neill, Mark Hills, Mike Gottlieb (Modified by David Porubsky)
}
