% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderContigs.R
\name{orderContigsTSP}
\alias{orderContigsTSP}
\title{Order contigs based on shared strand states}
\usage{
orderContigsTSP(
  contig.states = NULL,
  dist.matrix = NULL,
  method = "nearest_insertion",
  trials = 1000,
  filt.cols = TRUE
)
}
\arguments{
\item{contig.states}{A \code{data.frame} of strand states per contig and per cell.}

\item{dist.matrix}{A symetric \code{matrix} of similarities for each contig pair.}

\item{method}{Set one of the method to solve TSP, default: 'cheapest_insertion'.}

\item{trials}{Number of random trials to solve TSP.}

\item{filt.cols}{If set to \code{TRUE}, will remove columns with the same strand-state across all contigs.}
}
\description{
This function takes \code{data.frame} of strand states per contig [rows] and per cell [columns]
and tries to order contigs based on shared strand states among all cells.
}
\author{
David Porubsky
}
