% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAssemblyStat}
\alias{plotAssemblyStat}
\title{Plot assembly statistics}
\usage{
plotAssemblyStat(infile = NULL, format = "bam")
}
\arguments{
\item{infile}{A file that contains assembled contigs in specific format.}

\item{format}{Use 'bam' for contigs aligned to the reference or 'fai' for fasta index file or 'GRanges' for \code{\link{GRanges-class}} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot assembly statistics
}
\author{
David Porubsky
}
