% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotClustAccuracy}
\alias{plotClustAccuracy}
\title{Plot accuracy of clustring in respect to known location of each PacBio read.}
\usage{
plotClustAccuracy(pVal.df = NULL, num.clusters = NULL,
  thresh = c(0.5, 0.6, 0.7, 0.8, 0.9, 0.99))
}
\arguments{
\item{pVal.df}{A \code{data.frame} of probability values.}

\item{num.clusters}{Number of cluster present in probability table.}

\item{thresh}{A \code{vector} of probability thresholds to obtain accuracy of clustering.}
}
\description{
Plot accuracy of clustring in respect to known location of each PacBio read.
}
\author{
David Porubsky
}
