% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotClusteredContigSizes}
\alias{plotClusteredContigSizes}
\title{Plot distribution of cluster sizes.}
\usage{
plotClusteredContigSizes(clustered.gr = NULL)
}
\arguments{
\item{clustered.gr}{A \code{\link{GRanges-class}} object with a contig region and their cluster assignment in the 'ID' metacolumn.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot distribution of cluster sizes.
}
\author{
David Porubsky
}
