% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotContigStrandStates}
\alias{plotContigStrandStates}
\title{Plot contig strand states per cell}
\usage{
plotContigStrandStates(
  contig.states = NULL,
  cluster.rows = FALSE,
  cluster.cols = FALSE,
  filt.cols = FALSE
)
}
\arguments{
\item{contig.states}{A \code{data.frame} of strand states per contig and per cell.}

\item{cluster.rows}{If set to \code{TRUE}, will order rows by hierarchical clustering.}

\item{cluster.cols}{If set to \code{TRUE}, will order columns by hierarchical clustering.}

\item{filt.cols}{If set to \code{TRUE}, will remove columns with the same strand-state across all contigs.}
}
\description{
This function takes \code{data.frame} of strand states per contig [rows] and per cell [columns]
and plots heatmap of order or unordered strand states.
}
\author{
David Porubsky
}
