% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCTGstat}
\alias{plotCTGstat}
\title{Plot total number and total length of all contigs analyzed and their number after filtering and clustering step.}
\usage{
plotCTGstat(ctg.stat = NULL)
}
\arguments{
\item{clustered.gr}{A \code{data.frame} object with columns containing contig names, their lengths and a unique index.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot total number and total length of all contigs analyzed and their number after filtering and clustering step.
}
\author{
David Porubsky
}
