% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotEMprobs}
\alias{plotEMprobs}
\title{Plot distribution of cluster assignment probabilities.}
\usage{
plotEMprobs(em.prob = NULL, prob.th = 0)
}
\arguments{
\item{em.prob}{A \code{matrix} of probability assignments per contig and per cluster.}

\item{prob.th}{Only regions with this probability and higher will be reported.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot distribution of cluster assignment probabilities.
}
\examples{
## Get example files
example.data <- system.file("extdata/data", "softClust_100K_5e+06bp_dynamic.RData", package = "SaaRclust")
EM.obj <- get(load(example.data))
## Plot size and orientation of each cluster
prob.plt <- plotEMprobs(em.prob = EM.obj$soft.pVal)

}
\author{
David Porubsky
}
