% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap of responsibilities of each PB reads for each cluster as a probability value.}
\usage{
plotHeatmap(pVal.df = NULL, colOrder = NULL, num.clusters = NULL)
}
\arguments{
\item{pVal.df}{A \code{data.frame} of probability values.}

\item{colOrder}{A \code{vector} of indices representing of cluster order on heatmap.}

\item{num.clusters}{Number of cluster present in probability table.}
}
\description{
Plot heatmap of responsibilities of each PB reads for each cluster as a probability value.
}
\author{
David Porubsky
}
