% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotQualMeasure}
\alias{plotQualMeasure}
\title{Plot data quality measures.}
\usage{
plotQualMeasure(summary.tab)
}
\arguments{
\item{summary.tab}{Imported data table.}
}
\description{
Takes imported data table and plots some relevant data quality measures.
}
\author{
David Porubsky
}

