% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotReadAlignments}
\alias{plotReadAlignments}
\title{Plot coverage of short reads mapped on top of PB reads}
\usage{
plotReadAlignments(minimap.tab = NULL)
}
\arguments{
\item{minimap.tab}{A \code{data.frame} of short read mappings per PacBio read in maf.}
}
\description{
Plot coverage of short reads mapped on top of PB reads
}
\author{
David Porubsky
}

