% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotReadMappingDist}
\alias{plotReadMappingDist}
\title{Plot distribution of short reads mapped on top of PB reads}
\usage{
plotReadMappingDist(count.list = NULL)
}
\arguments{
\item{count.list}{A \code{list} of short read mappings per library.}
}
\description{
Plot distribution of short reads mapped on top of PB reads
}
\author{
David Porubsky
}

