% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotStrandStateZscore}
\alias{plotStrandStateZscore}
\title{Plot distribution of zscores reflecting abundance of WC states per cluster.}
\usage{
plotStrandStateZscore(zscores = NULL)
}
\arguments{
\item{zscores}{A \code{vector} of zscore values that reflect abundance of WC states per cluster and per Strand-seq library.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot distribution of zscores reflecting abundance of WC states per cluster.
}
\author{
David Porubsky
}
