% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{randomTheta}
\alias{randomTheta}
\title{Simulate random theta estimates for cell type}
\usage{
randomTheta(num.cells = 100, num.clusters = 44)
}
\arguments{
\item{num.cells}{Number of cells used in the analysis.}

\item{num.clusters}{Expected number of clusters. (for 22 autosomes == 44 clusters)}
}
\description{
Function to simulate random theta values in order to initialize EM algorithm.
}
\author{
David Porubsky
}
