% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionsToFasta.R
\name{regions2FASTA}
\alias{regions2FASTA}
\title{Export FASTA sequences from a set of Genomic Ranges.}
\usage{
regions2FASTA(
  gr,
  bsgenome = NULL,
  asm.fasta = NULL,
  expand = 0,
  fasta.save = NULL
)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object of genomic regions to extract genomic sequence from.}

\item{bsgenome}{A \code{\link{BSgenome-class}} object of reference genome to get the genomic sequence from.}

\item{asm.fasta}{An assembly FASTA file to extract DNA sequence determined by 'gr' parameter.}

\item{expand}{Expand ends of each genomic region by this length (in bp).}

\item{fasta.save}{A path to a filename where to store final FASTA file.}
}
\description{
This function takes a \code{\link{GRanges-class}} object and extracts a genomic sequence 
from these regions either from an original range or a range expanded on each side by
defined number of bases.
}
\author{
David Porubsky
}
