% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reportMisAsmCTGs}
\alias{reportMisAsmCTGs}
\title{Report summary of misassembled contigs/scaffolds}
\usage{
reportMisAsmCTGs(gr)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}
}
\value{
A \code{data.frame} object.
}
\description{
This function takes a \code{\link{GRanges-class}} object that contains putatively
misassembled contigs/scaffolds and report if this misassembly is likely caused by
an misorient, chimerism or both.
}
\author{
David Porubsky
}
