% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumped_functions.R
\name{rescaleTheta}
\alias{rescaleTheta}
\title{Rescale theta values for WC cell type}
\usage{
rescaleTheta(theta.l)
}
\arguments{
\item{theta.l}{A \code{list} of estmated theta values for every single cell.}
}
\description{
Set WC probs based on WW and CC probs distribution (if ratio of WW and CC == 1 then WC region with high prob)
}
\author{
David Porubsky
}
