% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{runSaaRclust}
\alias{runSaaRclust}
\title{Wrapper function to run saarclust pipeline for a given number of PB reads.}
\usage{
runSaaRclust(minimap.file = NULL, num.clusters = 44, EM.iter = 50,
  chunk.size = 10000)
}
\arguments{
\item{minimap.file}{A path to the minimap file to load.}

\item{num.clusters}{Expected number of clusters. (for 22 autosomes == 44 clusters)}

\item{EM.iter}{Number of iteration to run EM for.}

\item{chunk.size}{Number of PB reads to precess.}
}
\description{
Wrapper function to run saarclust pipeline for a given number of PB reads.
}
\author{
David Porubsky, Maryam Ghareghani
}

