% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saaRclust.R
\name{saarclust}
\alias{saarclust}
\title{EM function}
\usage{
saarclust(tab.l, theta.l = NULL, pi.param = NULL, num.iter = 100,
  raw.counts = NULL)
}
\arguments{
\item{tab.l}{A \code{list} of PB alignmetns separated per cell.}

\item{pi.param}{A \code{vector} of estimated sizes of each cluster based on initial hard clustering.}

\item{num.iter}{Set number of iteration to EM algorithm.}

\item{raw.counts}{Counts of +/- reads for evrry PB read.}

\item{theta.param}{A \code{list} of estimated cell types for each cluster and each cell.}
}
\value{
A \code{list} of various exported results [pVal, pVal.logL, log.l, theta.param, pi.param].
}
\description{
This function performs basic steps of EM algorithm.
}
\author{
David Porubsky, Maryam Ghareghani
}

