% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffoldDenovoAssembly.R
\name{scaffoldDenovoAssembly}
\alias{scaffoldDenovoAssembly}
\title{Main function of the \pkg{\link{SaaRclust}} package to order and orient contigs of denovo genome assembly.}
\usage{
scaffoldDenovoAssembly(bamfolder, outputfolder, configfile = NULL,
  min.contig.size = 1e+05, pairedEndReads = TRUE, bin.size = 1e+05,
  step.size = NULL, bin.method = "fixed", store.data.obj = TRUE,
  reuse.data.obj = FALSE, num.clusters = 100, alpha = 0.1,
  best.prob = 1, prob.th = 0, ord.method = "TSP",
  assembly.fasta = NULL, concat.fasta = TRUE, z.limit = 3.29,
  remove.always.WC = FALSE, mask.collapses = FALSE)
}
\arguments{
\item{bamfolder}{A folder containing BAM files with Strand-seq reads aligned to denovo assembly.}

\item{outputfolder}{Folder to output the results. If it does not exist it will be created.}

\item{configfile}{A file specifying the parameters of this function (without \code{bamfolder}, \code{outputfolder} and \code{configfile}). 
If a \code{configfile} is specified, it will take priority over the command line parameters.}

\item{min.contig.size}{A minimal contig size to be processed (default: 100kb).}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{bin.size}{A length of a bin to count reads in.}

\item{bin.method}{One of the 'fixed' or 'dynamic' binning method.}

\item{store.data.obj}{A logical indicating whether or not intermediate Rdata objects should be stored.}

\item{reuse.data.obj}{A logical indicating whether or not existing files in \code{outputfolder} should be reused.}

\item{num.clusters}{Expected number of clusters. (for 22 autosomes == 44 clusters)}

\item{alpha}{Estimated level of background in Strand-seq reads.}

\item{best.prob}{Set number of highest probability clusters reported per genomic region.}

\item{prob.th}{Only regions with this probability and higher will be reported.}

\item{ord.method}{A method ('TSP' or 'greedy') used to order contigs within a cluster (default: 'TSP')}

\item{assembly.fasta}{An assembly FASTA file to extract DNA sequence determined by clustered.gr.}

\item{concat.fasta}{Set to \code{TRUE} if you want to concatenate FASTA sequences within a cluster by 100 N's.}

\item{z.limit}{Connect clusters with z-score equal or above this limit.}

\item{remove.always.WC}{Set to \code{TRUE} if the cluster with majority of WC states should be removed.}
}
\value{
A \code{\link{GRanges-class}} object ???
}
\description{
This function is an easy-to-use wrapper to take Strand-seq reads aligned to the denovo genome assembly
and cluster each contig into chromosomal scaffold. In addition contigs are oriented and ordered within each
cluster/scaffold.
}
\author{
David Porubsky
}
