% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFASTA.R
\name{splitFASTA}
\alias{splitFASTA}
\title{A function to split FASTA file into a separate files one per FASTA sequence.}
\usage{
splitFASTA(fasta.file = NULL, outputfolder = NULL)
}
\arguments{
\item{fasta.file}{A FASTA file to be divided into individual FASTA sequences.}

\item{outputfolder}{A path to a folder where the output will be written.}
}
\description{
A function to split FASTA file into a separate files one per FASTA sequence.
}
\author{
David Porubsky
}
