% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncClusterDir.R
\name{syncClusterDir}
\alias{syncClusterDir}
\title{Re-orient contigs based on shared strand states}
\usage{
syncClusterDir(contig.states)
}
\arguments{
\item{contig.states}{A \code{data.frame} of strand states per contig and per cell.}
}
\description{
This function tries to find set of contigs inverted in respect to each other in every single cell
and flips directionality of such 'misoriented' contigs.
}
\author{
David Porubsky
}
