% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumped_functions.R
\name{WandCratiosFeatures}
\alias{WandCratiosFeatures}
\title{Get the feature vector based on both W and C ratios}
\usage{
WandCratiosFeatures(counts.l = NULL)
}
\arguments{
\item{counts.l}{A \code{list} of directional read counts per PB read per library.}
}
\value{
A \code{list} of estimated theta values for every cluster and cell.
}
\description{
This function expects output from custom minimap test dataset that contains original locations of mapped reads in the genome.
}
\author{
Maryam Ghareghani
}
