% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardClust.R
\name{WratiosFeatures}
\alias{WratiosFeatures}
\title{Get the feature vector based on the W ratios}
\usage{
WratiosFeatures(counts.l = NULL)
}
\arguments{
\item{counts.l}{A \code{list} of directional read counts per PB read per library.}
}
\value{
A \code{list} of estimated theta values for every cluster and cell.
}
\description{
This function expects output from custom minimap test dataset that contains original locations of mapped reads in the genome.
}
\author{
Maryam Ghareghani
}

