% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwConfig.R
\name{writeConfig}
\alias{writeConfig}
\title{Write SaaRclust configuration file}
\usage{
writeConfig(config, configfile)
}
\arguments{
\item{config}{A list structure with parameter values. Each entry will be written in one line.}

\item{configfile}{Filename of the outputfile.}
}
\value{
\code{NULL}
}
\description{
Write a SaaRclust configuration file from a list structure.
}
\author{
Aaron Taudt, David Porubsky
}
