% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{exportClusteredReads}
\alias{exportClusteredReads}
\title{Print names of long reads into separate files based on soft clustering assignment.}
\usage{
exportClusteredReads(inputfolder = NULL, prob.th = NULL,
  minLib = NULL)
}
\arguments{
\item{inputfolder}{Path to the data analysis folder}

\item{prob.th}{Filter out long reads with max probability below this threshold}

\item{minLib}{Filter out long reads with number of StrandS libraries being represented below this threshold}
}
\description{
Print names of long reads into separate files based on soft clustering assignment.
}
\author{
David Porubsky
}
