% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusterPartners.R
\name{findClusterPartners}
\alias{findClusterPartners}
\title{Get pairs of clusters coming from the same chromosome that differs in directionality.}
\usage{
findClusterPartners(theta.param = NULL)
}
\arguments{
\item{theta.param}{A \code{list} of estimated cell types for each cluster and each cell.}
}
\value{
A \code{matrix} of pairs of clusters IDs that belong to the same chromosome.
}
\description{
This function solves problem of finding the best clustering partners using linear programing function.
}
\author{
David Porubsky
}
