% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProbs.R
\name{gammaFunction}
\alias{gammaFunction}
\title{Calculate gamma function of EM algorithm}
\usage{
gammaFunction(clust.prob = NULL, pi.scaled = NULL, cellNum = NULL,
  log.scale = FALSE)
}
\arguments{
\item{clust.prob}{A \code{list} of binomial probabilities (multiplied by theta & pi parameter) of each PB read in every possible cluster}

\item{pi.scaled}{Paramter estimate of the size of each cluster scaled by cell number}

\item{cellNum}{Number of Strand-seq cells used for clustering.}

\item{log.scale}{Set to \code{TRUE} if probabilty should be calculated in a log scale. Recommended for high reads count 
numbers expected for larger genomic regions such as contigs.}
}
\value{
A \code{matrix} of unnormalized probalities of cell states per cluster.
}
\description{
Takes binomial probabilites for every cluster and update these probabilies given the assigment of Pacbio reads to clusters.
}
\author{
David Porubsky, Maryam Ghareghani
}
