% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{getQualMeasure}
\alias{getQualMeasure}
\title{Obtain summary measures of minimap output.}
\usage{
getQualMeasure(inputData)
}
\arguments{
\item{inputData}{A \code{data.frame} loaded by \code{\link[SaaRclust]{importTestData}}}
}
\description{
Takes imported data table and export relevant data quality measures.
}
\author{
David Porubsky
}
