% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getRepresentativeAlignments}
\alias{getRepresentativeAlignments}
\title{Import representative PacBio alignments.}
\usage{
getRepresentativeAlignments(inputfolder = NULL, numAlignments = 30000,
  quantileSSreads = c(0.2, 0.8), minSSlibs = c(20, 25))
}
\arguments{
\item{inputfolder}{A folder name where minimap files are stored.}

\item{numAlignments}{A required number of representative PacBio alignments.}

\item{quantileSSreads}{A quantile range for number of SSreads mapped to PB read. (default: 0.4-0.9)}

\item{minSSlibs}{A range for the minimal and maximal number of StrandS libs being represented per PB read.}
}
\value{
A \code{data.frame}.
}
\description{
This function selects the best representative PacBio alignements from multiple chunks
in order to get the best estimates of theta values using Kmeans.
}
\author{
David Porubsky
}

