% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardClust.R
\name{hardClust}
\alias{hardClust}
\title{Hard clustering using k-means}
\usage{
hardClust(tab.l = NULL, num.clusters = NULL, alpha = 0.1)
}
\arguments{
\item{tab.l}{A \code{list} of PB alignmetns separated per cell.}

\item{num.clusters}{Expected number of clusters. (for 22 autosomes == 44 clusters)}

\item{alpha}{Estimated level of background Strand-seq reads.}
}
\value{
A \code{list} of estimated theta values for every cluster and cell.
}
\description{
This function expects output from custom minimap test dataset that contains original locations of mapped reads in the genome.
}
\author{
David Porubsky
}

