% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardClust.R
\name{hardClust}
\alias{hardClust}
\title{Hard clustering using k-means}
\usage{
hardClust(counts.l = NULL, num.clusters = NULL, nstart = 10,
  iter.max = 10)
}
\arguments{
\item{counts.l}{A \code{list} of directional read counts per long read/contig per library.}

\item{num.clusters}{Desired number of clusters or cluster centers (k-means clustering).}

\item{nstart}{Number of random sets to choose cluster centers (k-means clustering).}

\item{iter.max}{A maximum number of allowed interations (k-means clustering).}
}
\value{
A \code{list} of estimated theta values for every cluster and cell.
}
\description{
This function expects output from custom minimap test dataset that contains original locations of mapped reads in the genome.
}
\examples{
## Get an example file
exampleFile <- system.file("extdata", "rawCounts_5e+06bp_dynamic.RData", package = "SaaRclust")
## Load BAM count table
counts.l <- get(load(exampleFile))
## Get hard clustering results
hardClust.ord <- hardClust(counts.l, num.clusters=100, nstart = 100)

}
\author{
David Porubsky
}
