% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuctions.R
\name{hardClustAccuracy}
\alias{hardClustAccuracy}
\title{Computes the performance of the hard clustering algorithm}
\usage{
hardClustAccuracy(hard.clust, classes, tab.filt, female = TRUE)
}
\arguments{
\item{hard.clust}{hard clustring result}

\item{tab.filt}{table of filtered read counts}

\item{female}{a binary argument showing the sex of the individual}
}
\description{
Computes the performance of the hard clustering algorithm
}
\author{
Maryam Ghareghani
}
