% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBins.R
\name{makeDynamicBins}
\alias{makeDynamicBins}
\title{Make dynamic-width bins}
\usage{
makeDynamicBins(bamfiles = NULL, bin.size = 1e+05, step.size = NULL,
  chromosomes = NULL, keep.small.chr = FALSE)
}
\arguments{
\item{bin.size}{A size of the genomic bin to split genome into.}

\item{step.size}{A size of the genomic interval to move each bin.}

\item{chromosomes}{A user defined set of chromosomes for binning (eg. 'chr1')}

\item{keep.small.chr}{Set to \code{TRUE} if chromosome/contigs smaller than the 'bin.size' should be kept.}
}
\value{
A \code{\link{GRanges-class}} object with fixed-width bins.
}
\description{
Make dynamic-width bins based on given bin and step size.
NOTE: Here 'bin.size' represents number of mappable positions in each bin !!!
}
\author{
David Porubsky
}
