% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderAndOrientClusters.R
\name{orderAndOrientClusters}
\alias{orderAndOrientClusters}
\title{Order and orient contigs within each cluster}
\usage{
orderAndOrientClusters(clustered.grl, split.pairs, ord.method = "TSP",
  alpha = 0.1, bin.size = bin.size, filename = NULL)
}
\arguments{
\item{clustered.grl}{A \code{\link{GRangesList-class}} object with cluster assignments for each contig.}

\item{split.pairs}{A \code{list} with a sets of cluster IDs that belong the same scaffold/chromosome.}

\item{ord.method}{A method ('TSP' or 'greedy') used to order contigs within a cluster (default: 'TSP')}

\item{alpha}{Estimated level of background in Strand-seq reads.}

\item{bin.size}{A length of a bin to count reads in.}

\item{filename}{A path to a file where clustered on ordered contigs should be stored.}
}
\description{
This function takes a \code{\link{GRangesList-class}} object and order and orient each contigs
within the cluster and reports new clustered/scaffolded FASTA file.
}
\author{
David Porubsky
}
