% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotClusteredContigs}
\alias{plotClusteredContigs}
\title{Plot genome-wide positions of clustered contigs}
\usage{
plotClusteredContigs(bedfile, min.mapq = 10, bsgenome = NULL,
  blacklist = NULL, report = "clustering", min.contig.size = NULL,
  reverse.x = FALSE)
}
\arguments{
\item{bedfile}{An aligned contigs to the reference sequence in bed format.}

\item{min.mapq}{Minimum mapping quality of a contig to the refrence sequence.}

\item{bsgenome}{A \code{\link{GBSgenome-class}} object to provide chromosome lengths for plotting.}

\item{blacklist}{A \code{\link{GRanges-class}} object of regions to be removed.}

\item{report}{Plot either 'clustering', 'ordering' or 'orienting' of the contigs. Default: 'clustering'.}

\item{min.contig.size}{Minimal contigs size to plot.}

\item{reverse.x}{Set to \code{TRUE} if x-axis should be horizontaly reversed.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot genome-wide positions of clustered contigs
}
\author{
David Porubsky
}
