% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotContigsGenomeWide}
\alias{plotContigsGenomeWide}
\title{Plot genome-wide positions of clustered contigs}
\usage{
plotContigsGenomeWide(gr = NULL, bsgenome = NULL, blacklist = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object with contig position and their cluster assignment in 'clust.ID' and 'group.ID' metacolumn.}

\item{bsgenome}{A \code{\link{GBSgenome-class}} object to provide chromosome lengths for plotting.}

\item{blacklist}{A \code{\link{GRanges-class}} object of regions to be removed from input gr.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot genome-wide positions of clustered contigs
}
\author{
David Porubsky
}
