% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDistanceMatrix}
\alias{plotDistanceMatrix}
\title{...}
\usage{
plotDistanceMatrix(dist.matrix, col.low = "chartreuse4",
  col.high = "cadetblue1")
}
\arguments{
\item{dist.matrix}{...}

\item{col.low}{...}

\item{col.high}{...}
}
\value{
A \code{ggplot} object.
}
\description{
...
}
\author{
David Porubsky
}
