% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotEMprobs}
\alias{plotEMprobs}
\title{Plot distribution of cluster assignment probabilities.}
\usage{
plotEMprobs(em.prob = NULL, prob.th = 0)
}
\arguments{
\item{em.prob}{A \code{matrix} of probability assignments per contig and per cluster.}

\item{prob.th}{Only regions with this probability and higher will be reported.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot distribution of cluster assignment probabilities.
}
\author{
David Porubsky
}
