% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotThetaEstimates}
\alias{plotThetaEstimates}
\title{Plot theta estimates resulting from EM algorithm.}
\usage{
plotThetaEstimates(theta.param = NULL, title = NULL)
}
\arguments{
\item{theta.param}{A \code{list} of estimated cell types for each cluster and each cell.}

\item{title}{A \code{character} to use as a title of the plot.}
}
\description{
Plot theta estimates resulting from EM algorithm.
}
\author{
David Porubsky
}
