% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{runSaaRclust}
\alias{runSaaRclust}
\title{Wrapper function to run saarclust pipeline.}
\usage{
runSaaRclust(inputfolder = NULL, outputfolder = "SaaRclust_results",
  num.clusters = 55, EM.iter = 100, alpha = 0.01, logL.th = 1,
  theta.constrain = FALSE, store.counts = FALSE, store.bestAlign = TRUE,
  skip.HC = FALSE, verbose = TRUE)
}
\arguments{
\item{inputfolder}{A folder name where minimap files are stored.}

\item{outputfolder}{A folder name to export to results.}

\item{num.clusters}{Expected number of clusters. (for 22 autosomes == 44 clusters)}

\item{EM.iter}{Number of iteration to run EM for.}

\item{alpha}{Estimated level of background Strand-seq reads.}

\item{store.counts}{Logical if to store raw read counts per PB read}

\item{verbose}{...}
}
\description{
Wrapper function to run saarclust pipeline.
}
\author{
David Porubsky, Maryam Ghareghani
}

