% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{runSaaRclust}
\alias{runSaaRclust}
\title{Wrapper function to run saarclust pipeline.}
\usage{
runSaaRclust(inputfolder = NULL, outputfolder = "SaaRclust_results",
  num.clusters = 54, EM.iter = 100, alpha = 0.01, minLib = 10,
  upperQ = 0.95, logL.th = 1, theta.constrain = FALSE,
  store.counts = FALSE, store.bestAlign = TRUE,
  numAlignments = 30000, HC.only = TRUE, verbose = TRUE,
  cellNum = NULL, log.scale = FALSE)
}
\arguments{
\item{inputfolder}{A folder name where minimap files are stored.}

\item{outputfolder}{A folder name to export the results.}

\item{num.clusters}{Expected number of clusters. (for 22 autosomes == 44 clusters)}

\item{EM.iter}{Number of iteration to run EM for.}

\item{alpha}{Estimated level of background in Strand-seq reads.}

\item{minLib}{Minimal number of StrandS libraries being represent per long PB read}

\item{upperQ}{Filter out given percentage of PacBio reads with the highest number of alignments.}

\item{logL.th}{Set the difference between objective function from the current and the previous interation for EM algorithm to converge.}

\item{theta.constrain}{Recalibrate theta values to meet expected distribution of W and C strand across Strand-seq libraries.}

\item{store.counts}{Logical if to store raw read counts per PB read}

\item{store.bestAlign}{Store best alignements in RData object.}

\item{numAlignments}{Required number of best PBvsSS alignmnets to selest for hard clustering.}

\item{HC.only}{Perform only hard clustering and skip the rest of the pipeline.}

\item{verbose}{Set to \code{TRUE} to print function messages.}

\item{cellNum}{specifies the number of single cells to be used in clustering}

\item{HC.input}{Filaname where hard clustering results are stored}
}
\description{
Wrapper function to run saarclust pipeline.
}
\author{
David Porubsky, Maryam Ghareghani
}
