% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaRclust.R
\name{SaaRclust}
\alias{SaaRclust}
\title{Wrapper function to run saarclust pipeline for a given number of PB reads.}
\usage{
SaaRclust(minimap.file = NULL, outputfolder = "SaaRclust_analysis",
  num.clusters = 44, EM.iter = 100, chunk.size = 20000, alpha = 0.1,
  logL.th = 1, theta.constrain = FALSE)
}
\arguments{
\item{minimap.file}{A path to the minimap file to load.}

\item{outputfolder}{A folder name to export to results.}

\item{num.clusters}{Expected number of clusters. (for 22 autosomes == 44 clusters)}

\item{EM.iter}{Number of iteration to run EM for.}

\item{chunk.size}{Number of PB reads to precess.}

\item{alpha}{Estimated level of background Strand-seq reads.}
}
\description{
Wrapper function to run saarclust pipeline for a given number of PB reads.
}
\author{
David Porubsky
}

