% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaRclust.R
\name{SaaRclust}
\alias{SaaRclust}
\title{Wrapper function to run saarclust pipeline for a given number of PB reads.}
\usage{
SaaRclust(minimap.file = NULL, outputfolder = "SaaRclust_results",
  num.clusters = 47, EM.iter = 100, alpha = 0.1, minLib = 10,
  upperQ = 0.95, theta.param = NULL, pi.param = NULL, logL.th = 1,
  theta.constrain = FALSE, store.counts = FALSE, HC.input = NULL,
  cellNum = NULL)
}
\arguments{
\item{minimap.file}{A path to the minimap file to load.}

\item{outputfolder}{A folder name to export to results.}

\item{num.clusters}{Expected number of clusters. (for 22 autosomes == 44 clusters)}

\item{EM.iter}{Number of iteration to run EM for.}

\item{alpha}{Estimated level of background Strand-seq reads.}

\item{minLib}{Minimal number of StrandS libraries being represent per long PB read}

\item{upperQ}{Filter out given percentage of PacBio reads with the highest number of alignments.}

\item{store.counts}{Logical if to store raw read counts per PB read}

\item{HC.input}{Filaname where hard clustering results are stored}

\item{cellNum}{specifies the number of single cells to be used in clustering}
}
\description{
Wrapper function to run saarclust pipeline for a given number of PB reads.
}
\author{
David Porubsky
}
