% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaaRclust.R
\name{SaaRclust}
\alias{SaaRclust}
\title{Wrapper function to run saarclust pipeline for a given number of long reads reads.}
\usage{
SaaRclust(
  minimap.file = NULL,
  outputfolder = "SaaRclust_results",
  num.clusters = 47,
  EM.iter = 100,
  alpha = 0.1,
  minLib = 10,
  upperQ = 0.95,
  theta.param = NULL,
  pi.param = NULL,
  logL.th = 1,
  theta.constrain = FALSE,
  store.counts = FALSE,
  HC.input = NULL,
  cellNum = NULL,
  log.scale = FALSE
)
}
\arguments{
\item{minimap.file}{A path to the minimap file to load.}

\item{outputfolder}{A folder name to export the results.}

\item{num.clusters}{Expected number of clusters. (for 22 autosomes == 44 clusters)}

\item{EM.iter}{Number of iteration to run EM for.}

\item{alpha}{Estimated level of background in Strand-seq reads.}

\item{minLib}{Minimal number of StrandS libraries being represent per long PB read}

\item{upperQ}{Filter out given percentage of PacBio reads with the highest number of alignments.}

\item{theta.param}{A \code{list} of estimated cell types for each single cell. (rows=clusters, cols=strand states)}

\item{pi.param}{A \code{vector} of estimated sizes of each cluster based on initial hard clustering.}

\item{logL.th}{Set the difference between objective function from the current and the previous interation for EM algorithm to converge.}

\item{theta.constrain}{Recalibrate theta values to meet expected distribution of W and C strand across Strand-seq libraries.}

\item{store.counts}{Logical if to store raw read counts per PB read}

\item{HC.input}{Filaname where hard clustering results are stored}

\item{cellNum}{Specifies the number of single cells to be used in clustering}

\item{log.scale}{Set to \code{TRUE} if probabilty should be calculated in a log scale. Recommended for high reads count 
numbers expected for larger genomic regions such as contigs.}
}
\description{
Wrapper function to run saarclust pipeline for a given number of long reads reads.
}
\author{
David Porubsky
}
