% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffoldDenovoAssembly.R
\name{scaffoldDenovoAssembly}
\alias{scaffoldDenovoAssembly}
\title{Main function of the \pkg{\link{SaaRclust}} package to order and orient contigs of denovo genome assembly.}
\usage{
scaffoldDenovoAssembly(
  bamfolder,
  outputfolder,
  configfile = NULL,
  min.mapq = 10,
  min.contig.size = 1e+05,
  min.region.to.order = 0,
  pairedEndReads = TRUE,
  bin.size = 1e+05,
  step.size = NULL,
  bin.method = "fixed",
  store.data.obj = TRUE,
  reuse.data.obj = FALSE,
  num.clusters = 100,
  desired.num.clusters = NULL,
  alpha = 0.1,
  best.prob = 1,
  prob.th = 0,
  ord.method = "TSP",
  assembly.fasta = NULL,
  concat.fasta = TRUE,
  z.limit = 3.29,
  remove.always.WC = FALSE,
  mask.regions = FALSE,
  eval.haploid = FALSE
)
}
\arguments{
\item{bamfolder}{A folder containing BAM files with Strand-seq reads aligned to denovo assembly.}

\item{outputfolder}{Folder to output the results. If it does not exist it will be created.}

\item{configfile}{A file specifying the parameters of this function (without \code{bamfolder}, \code{outputfolder} and \code{configfile}). 
If a \code{configfile} is specified, it will take priority over the command line parameters.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{min.contig.size}{A minimal contig size to be processed (default: 100kb).}

\item{min.region.to.order}{A minimum size region to try to order within a cluster.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{bin.size}{A length of a bin to count reads in.}

\item{step.size}{A size of the genomic interval to move each bin.}

\item{bin.method}{One of the 'fixed' or 'dynamic' binning method.}

\item{store.data.obj}{A logical indicating whether or not intermediate Rdata objects should be stored.}

\item{reuse.data.obj}{A logical indicating whether or not existing files in \code{outputfolder} should be reused.}

\item{num.clusters}{Desired number of clusters or cluster centers (k-means clustering).}

\item{desired.num.clusters}{Desired number of clusters after cluster merging step.}

\item{alpha}{Estimated level of background in Strand-seq reads.}

\item{best.prob}{Set number of highest probability clusters reported per genomic region.}

\item{prob.th}{Only regions with this probability and higher will be reported.}

\item{ord.method}{A method ('TSP' or 'greedy') used to order contigs within a cluster (default: 'TSP')}

\item{assembly.fasta}{An assembly FASTA file to extract DNA sequence determined by clustered.gr.}

\item{concat.fasta}{Set to \code{TRUE} if you want to concatenate FASTA sequences within a cluster by 100 N's.}

\item{z.limit}{Connect clusters with z-score equal or above this limit.}

\item{remove.always.WC}{Set to \code{TRUE} if the cluster with majority of WC states should be removed.}

\item{mask.regions}{Set to \code{TRUE} if regions that appear as WC in majority of cells and low coverage regions should be masked.}

\item{eval.haploid}{If set to \code{TRUE} likely haploid clusters won't be merged based on WC state probabilities. This option
helps to properly resolve sex chromosomes in males.}
}
\description{
This function is an easy-to-use wrapper to take Strand-seq reads aligned to the denovo genome assembly
and cluster each contig into chromosomal scaffold. In addition contigs are oriented and ordered within each
cluster/scaffold.
}
\examples{
\dontrun{
## Required parameters to run SaaRclust on BAM files stored "bam-data-folder" using default settings.
## To export clustred FASTA file, an original FASTA used in BAM alignments has to be submitted as 'assembly.fasta'.
scaffoldDenovoAssembly(bamfolder="bam-data-folder", outputfolder="saarclust-output-folder")}

}
\author{
David Porubsky
}
