% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string2GRanges}
\alias{string2GRanges}
\title{Convert strings in a data.frame object into a \code{\link{GRanges-class}} object.}
\usage{
string2GRanges(region.string = NULL)
}
\arguments{
\item{region.string}{A \code{vector} object}
}
\value{
A \code{\link{GRanges-class}} object with an extra metacolumn group.ID.
}
\description{
Convert strings in a data.frame object into a \code{\link{GRanges-class}} object.
}
\examples{
## An example string to convert into \\code{\\link{GRanges-class}} object.
genomic.location.str <- 'chr4:128840654-143665255'
gr <- string2GRanges(region.string = genomic.location.str)

}
\author{
David Porubsky
}
